requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.11.4
google-auth==2.23.3
python-dotenv==1.0.0
gunicorn==21.2.0
google-cloud-geminidataanalytics==0.3.0